Put a player's starting cash $$ into PokerLib
Use it to initialize both the handscreen and the table. This will get
overridden quickly once a hand is dealt and ante is subtracted, but
there's enough time for the value to flicker on the client screen and it
looks like a bug if it goes from 100k to 49k.