On a fold, make the players' cards invisible
The protocol is that the server always acknowledges every action the
player requests and only then does the player change anything. So ACK
the Fold action, and upon that ACK make the cards invisible. You won't
be in the round anymore so the ACTION_ACKNOWLEDGE callback won't be
invoked again until you are no longer folded.